+++ types.d.ts
@@ -1,5 +1,7 @@
 declare global {
   type TeqFw_Di_Config_NamespaceRegistry = import("./src/Config/NamespaceRegistry.mjs").default;
+  type TeqFw_Di_Config_PackageRegistry = import("./src/Config/PackageRegistry.mjs").default;
+  type TeqFw_Di_Config_PackageRegistry_Record = Readonly<{name: string, rootAbs: string, rootReal: string, packageJson: Readonly<Record<string, unknown>>}>;
   type TeqFw_Di_Container = import("./src/Container.mjs").default;
   type TeqFw_Di_Container_Instantiate = import("./src/Container/Instantiate.mjs").default;
   type TeqFw_Di_Container_Lifecycle = import("./src/Container/Lifecycle.mjs").default;